Rest node server tests done

+++ src/madsci_node_module/madsci/node_module/rest_node_module.py
@@ -7,7 +7,6 @@ import time
 from contextlib import asynccontextmanager
 from multiprocessing import Process
 from pathlib import Path, PureWindowsPath
-from threading import Thread
 from typing import Any, Optional, Union
 from zipfile import ZipFile
 
@@ -15,7 +14,11 @@ from fastapi.applications import FastAPI
 from fastapi.datastructures import UploadFile
 from fastapi.routing import APIRouter
 from madsci.client.node.rest_node_client import RestNodeClient
-from madsci.common.types.action_types import ActionRequest, ActionResult, ActionStatus
+from madsci.common.types.action_types import (
+    ActionRequest,
+    ActionResult,
+    ActionStatus,
+)
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.base_types import Error, new_ulid_str
 from madsci.common.types.event_types import Event
@@ -178,9 +181,11 @@ class RestNode(AbstractNode):
             )
         return ActionResult.model_validate(action_response)
 
-    def get_action_history(self) -> list[str]:
-        """Get the action history of the node."""
-        return super().get_action_history()
+    def get_action_history(
+        self, action_id: Optional[str] = None
+    ) -> dict[str, list[ActionResult]]:
+        """Get the action history of the node, or of a specific action."""
+        return super().get_action_history(action_id)
 
     def get_status(self) -> NodeStatus:
         """Get the status of the node."""
@@ -194,7 +199,7 @@ class RestNode(AbstractNode):
         """Get the state of the node."""
         return super().get_state()
 
-    def get_log(self) -> list[Event]:
+    def get_log(self) -> dict[str, Event]:
         """Get the log of the node"""
         return super().get_log()
 
@@ -215,8 +220,9 @@ class RestNode(AbstractNode):
         try:
             self.restart_flag = True  # * Restart the REST server
             self.shutdown_handler()
-            self.startup_handler(self.config)
+            self._startup()
         except Exception as exception:
+            self._exception_handler(exception)
             return AdminCommandResponse(
                 success=False,
                 errors=[Error.from_exception(exception)],
@@ -235,7 +241,11 @@ class RestNode(AbstractNode):
             def shutdown_server() -> None:
                 """Shutdown the REST server."""
                 time.sleep(2)
-                self.rest_server_process.terminate()
+                if (
+                    hasattr(self, "rest_server_process")
+                    and self.rest_server_process is not None
+                ):
+                    self.rest_server_process.terminate()
                 self.exit_flag = True
 
             shutdown_server()
@@ -274,37 +284,14 @@ class RestNode(AbstractNode):
                 if self.restart_flag:
                     self.rest_server_process.terminate()
                     self.restart_flag = False
-                    self._start_rest_api()
+                    self._start_rest_api(testing=testing)
                     break
                 if self.exit_flag:
                     break
 
-    def _startup_thread(self) -> None:
-        """The startup thread for the REST API."""
-        try:
-            # * Create a clean status and mark the node as initializing
-            self.node_status.initializing = True
-            self.node_status.errored = False
-            self.node_status.locked = False
-            self.node_status.paused = False
-            self.startup_handler()
-        except Exception as exception:
-            # * Handle any exceptions that occurred during startup
-            self._exception_handler(exception)
-            self.node_status.errored = True
-        finally:
-            # * Mark the node as no longer initializing
-            self.logger.log(f"Startup complete for node {self.node_info.node_name}.")
-            self.node_status.initializing = False
-
     @asynccontextmanager
     async def _lifespan(self, app: FastAPI):  # noqa: ANN202, ARG002
         """The lifespan of the REST API."""
-        # * Run startup on a separate thread so it doesn't block the rest server from starting
-        # * (node won't accept actions until startup is complete)
-        Thread(target=self._startup_thread, daemon=True).start()
-        self._loop_handler()
-
         yield
 
         try:
@@ -338,6 +325,11 @@ class RestNode(AbstractNode):
             "/admin/{admin_command}",
             self.run_admin_command,
             methods=["POST"],
+        )
+        self.router.add_api_route(
+            "/log",
+            self.get_log,
+            methods=["GET"],
         )
         self.rest_api.include_router(self.router)
